add kmeans_helper.hpp to store reusable functions
- add kmeans_helper.hpp to store reusable functions
- use std::pair in function *get_nearest _center*
